refactor(router): extract the task operations into a child module
Phase 2 of #1256, taking the group the issue names as the cleanest first
extraction. `router.rs` goes 5251 to 4805 lines; `router/task_ops.rs` holds
the 12 methods.

The SEP-2663 task surface moves together because it is used together:
`classify_absent_task`, `require_negotiated_tasks`, `authorize_task`,
`final_get_task`, `detailed_task`, `park_task_for_input`, the four live-task
registry methods, `resume_task`, and `notify_task_state`.

An `impl McpRouter` block in a child module, so neither the type nor its API
changed. The methods were private and are now `pub(super)`, which is what a
child module needs to stay reachable from `router.rs` and is the narrowest
visibility that compiles; nothing outside `router` can see them.

The module is `task_ops` rather than `tasks` so it is not read as
`crate::tasks`, which holds the protocol types these methods build.

Test count is the guard, as in phase 1: 1030 lib tests before and after. The
extraction asserted both block boundaries against the doc comment that opens
them and the item that follows, and it caught its own off-by-one on the
second block before writing anything.

Refs #1256
